Save to tmp directory on non-chief workers in model_training_utils
In a multi worker set up saving is done on each worker. If they're saving to the same location, e.g. GCS, there will be conflicts. With this change we save to temporary directory on non-chief workers.

Note that, there may be synchronization in saving that needs all workers to participate, so we cannot only save on one worker.

PiperOrigin-RevId: 300141152